implements knightMoves function
This function receives two coordinates and returns the shortest path a knight can do to get from the start coordinates to the end. It uses a Breadth first search algorythm to navigate through the moves until it finds the end.
The way the path is saved, is by adding a reference to the previous move for each move visitated in a Map object, and then, after finding the end, it just has to navigate up the parents until it meets the start.